backend/requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.0
python-dotenv>=1.0.0

# Database
pymongo>=4.5.0

# AI & ML
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0

# Document Processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10

# WSGI Server
gunicorn>=21.2.0
whitenoise>=6.5.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
